nnas: renamed nnid service to nnas
