+++ src/server.ts
@@ -16,7 +16,7 @@ import { LOG_INFO, LOG_SUCCESS, LOG_WARN } from '@/logger';
 
 import conntest from '@/services/conntest';
 import cbvc from '@/services/cbvc';
-import nnid from '@/services/nnid';
+import nnas from '@/services/nnas';
 import nasc from '@/services/nasc';
 import datastore from '@/services/datastore';
 import api from '@/services/api';
@@ -41,7 +41,7 @@ app.use(xmlparser);
 // import the servers into one
 app.use(conntest);
 app.use(cbvc);
-app.use(nnid);
+app.use(nnas);
 app.use(nasc);
 app.use(datastore);
 app.use(api);
