+++ src/services/nnas/routes/devices.ts
