+++ src/types/services/nnas/region-languages.ts
@@ -1,0 +1,5 @@
+import { RegionTimezones } from '@/types/services/nnas/region-timezones';
+
+export interface RegionLanguages {
+	[myKey: string]: RegionTimezones
+}
