--- src/types/services/nnid/region-languages.ts
@@ -1,5 +1,0 @@
-import { RegionTimezones } from '@/types/services/nnid/region-timezones';
-
-export interface RegionLanguages {
-	[myKey: string]: RegionTimezones
-}
